GH-336 add ABS files, create basic mapping [skip ci]

+++ qendpoint-core/src/main/java/com/the_qa_company/qendpoint/core/storage/QEPDataset.java
@@ -5,6 +5,7 @@ import com.the_qa_company.qendpoint.core.hdt.HDT;
 import java.io.Closeable;
 import java.io.IOException;
 import java.nio.file.Path;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
 /**
@@ -15,8 +16,8 @@ import java.util.concurrent.atomic.AtomicLong;
  * @param dataset loaded dataset
  * @param uid     uid of the dataset for internal use
  */
-record QEPDataset(String id, Path path, HDT dataset, long uid) implements Closeable {
-	private static final AtomicLong DATASET_UID_FETCHER = new AtomicLong();
+record QEPDataset(String id, Path path, HDT dataset, int uid) implements Closeable {
+	private static final AtomicInteger DATASET_UID_FETCHER = new AtomicInteger();
 
 	public QEPDataset(String id, Path path, HDT dataset) {
 		this(id, path, dataset, DATASET_UID_FETCHER.incrementAndGet());
